Remove `state` parameter from `MappedNavigatableSource.build` method
Since the base class can not enforce that the parameter is marked as `final`, it's too easy to accidentally assign to the "local" `state` instead of `this.state` when one wants to initiate a state update.

This has the usual drawback though, that one must use `final state = state;` for complex type or non-`null` assertions and usages. But that is also a known issue with most other fields, and seems better than loosing updates.